fixed order

+++ src/main/java/com/kasia/services/products/StatisticsService.java
@@ -1,9 +1,8 @@
 package com.kasia.services.products;
 
-
-import java.util.HashMap;
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface StatisticsService {
-    Map<String, Double> averagePriceMap();
+    Map<String, BigDecimal> averagePriceMap();
 }
